+++ asabiris/exception_manager/api_exception_manager.py
@@ -1,4 +1,4 @@
-from asabiris.exception_manager_abc import ExceptionManager
+from asabiris.exception_manager.exception_manager_abc import ExceptionManager
 
 
 class APIExceptionManager(ExceptionManager):
